SyncDNS will now write the dns records from the blobstore
- Located at /var/vcap/instances/dns/records.json

[#136927903](https://www.pivotaltracker.com/story/show/136927903)

+++ settings/settings.go
@@ -202,7 +202,7 @@ type BoshEnv struct {
 }
 
 type DNSRecords struct {
-	Version uint32      `json:"Version"`
+	Version uint64      `json:"Version"`
 	Records [][2]string `json:"records"`
 }
